requirements.txt
fastapi
uvicorn
qdrant-client
openai
psycopg2-binary
asyncpg
python-dotenv
pydantic
pydantic-settings
httpx
aiohttp
tenacity